Two figures uploaded
Figure 1 describes the most frequent words appeared in all the reviews collected from Glassdoor. The top part of the figure shows the words frequently appeared in positive reviews; the bottom part of the figure shows the words frequently appeared in negative reviews. 

Figure 2 describes the comparison between different companies. It shows the words (or characteristics) appeared in the positive reviews categorized by different companies. This figure is an indication of the most attractive parts of each company. Such kind of results can be used to build a recommendation system for job applicants.